api/requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-environ>=0.11,<1.0
